[WIP] DuckyScript plugin
Allows a DAP to behave like a USB Rubber Ducky by executing DuckyScript
files,
which is a BASIC-like language with facilities for sending keystrokes.

A "crash course" on the language can be found at
http://71.71.39.6/ducky.txt

This plugin also implements several extensions to vanilla DuckyScript,
some of
which are listed below (see source for complete documentation).

- variables (multicharacter names via hash map)
- control flow (JUMP/GOTO, IF) with labels
- expression parsing when a number is needed
- non-decimal bases
- bytecode compilation (on computer only)

The compiler and related tools can be found here:
https://github.com/theunamedguy/ducky

NOTE: Changes some entries in the HID usage tables, but doesn't break
anything.

TODO:

- Finish up manual entry
- Further documentation?
- More optimization?
- Arrays (how?)
- TEST, TEST, TEST!
   - Some features have only been marginally tested and MIGHT NOT WORK
CORRECTLY
- FASTER TYPING!!!
   - Currently averages ~96 keys/sec
   - Theoretical maximum: 500 keys/sec
   - Of interest: G#1223 (http://gerrit.rockbox.org/r/1223/)

Change-Id: I149ecd08ba124f8b8637a002ee2652dd54a44e58