fuze+: don't block system on backlight change
The fuze+ backlight chip is stupid and requires a 600us pulse before sending
a new value. This delay is very long, especially when doing backlight fading
where we will easily accumulate this delay many times. For now, the code will
schedule a DPC and block the caller, thus yielding and scheduling another
thread instead of doing a busy wait.

Change-Id: Idc3821ff41214f706e263e792a95e10e07ae560d