Make HID Mouse more controllable and accurate
Uses a sublinear acceleration that makes it easy to hit the intended
target and has 1-pixel initial movement. IMO the constant chosen gives
as much control as could reasonably be expected from moving a mouse
cursor with keys-- not too slow for short distances but not so fast
you badly overshoot when moving a long distance.

Change-Id: I6160fb996e0607e94e977deb64c918210316c31e